summarizer_macos/pyproject.toml
[project]
name = "summarizer-macos"
version = "0.1.0"
description = "MLX based Gemini 4 summarizer for macOS"
requires-python = ">=3.13"
dependencies = [
    "mlx-vlm>=0.4.3",
    "pyyaml",
]

[project.scripts]
summarizer-macos = "summarizer_macos:main"
